[Style 🎨] 에스터리스크를 after 가상속성으로 대체
