+++ src/components/Input/components/InputError.tsx
@@ -1,7 +1,15 @@
-// @ts-nocheck
-import * as S from "../StyledInput.js"
+import { PropsWithChildren } from "react"
 
-const InputError = ({ children }) => {
-  return <S.InputWarning>{children}</S.InputWarning>
+import styled from "styled-components"
+
+import theme from "@styles/theme"
+
+const InputError = ({ children }: PropsWithChildren) => {
+  return <Warning>{children}</Warning>
 }
 export default InputError
+
+const Warning = styled.span`
+  color: ${theme.Error};
+  font-size: 16px;
+`
