+++ src/components/Input/components/InputInput.tsx
@@ -1,15 +1,16 @@
 import { InputHTMLAttributes } from "react"
 
-import * as S from "../StyledInput.js"
+import styled from "styled-components"
+
+import theme, { fonts } from "@styles/theme.js"
 
 interface InputInputProps extends InputHTMLAttributes<HTMLInputElement> {
-  // any 타입 고치기
   props: any
 }
 
 const InputInput = ({ props, ...rest }: InputInputProps) => {
   return (
-    <S.Input
+    <Input
       {...props}
       {...rest}
     />
@@ -17,3 +18,25 @@ const InputInput = ({ props, ...rest }: InputInputProps) => {
 }
 
 export default InputInput
+
+const getBorderStyle = ($isError: boolean, $isDirty: boolean) => {
+  if (!$isDirty) {
+    return `1px solid ${theme.Netural500}`
+  }
+  return $isError ? `2px solid ${theme.Error}` : `2px solid ${theme.Brand600}`
+}
+
+export const Input = styled.input<{
+  $isError: boolean
+  $isDirty: boolean
+}>`
+  &::-webkit-input-placeholder {
+    color: ${theme.Netural500};
+    ${fonts.b4}
+  }
+  background: ${theme.Netural0};
+  border-radius: 10px;
+  padding: 16px;
+  width: 100%;
+  border: ${({ $isError, $isDirty }) => getBorderStyle($isError, $isDirty)};
+`
